Replace FROM_DEPLOYMENTCONFIG with APP_EXIT and default to false
The loop forever behavior is now controlled with the APP_EXIT
env var which is defaulted to "false" in start.sh if it is not
set. The APP_EXIT env var is set to "true" in the job template
but not exposed.  It is set to "false" in the deploymentconfig
templates and exposed so that it may be overridden.

This will allow the default behavior from "oc new-app" to be
application pods that never exit.  This behavior may be
overridden by passing "-e APP_EXIT=true" to oc new-app
(for a streaming app for example which should restart if
it ever exists)