Replace FROM_DEPLOYMENTCONFIG with APP_EXIT and default to false (#34)
* Replace FROM_DEPLOYMENTCONFIG with APP_EXIT and default to false

The loop forever behavior is now controlled with the APP_EXIT
env var which is defaulted to "false" in start.sh if it is not
set. The APP_EXIT env var is set to "true" in the job template
but not exposed.  It is set to "false" in the deploymentconfig
templates and exposed so that it may be overridden.

This will allow the default behavior from "oc new-app" to be
application pods that never exit.  This behavior may be
overridden by passing "-e APP_EXIT=true" to oc new-app
(for a streaming app for example which should restart if
it ever exists)

* Add note on APP_EXIT to s2i/bin/usage

Primarily for use when running in a deploymentconfig but
document it anyway.

* Always rely on the pod environment for the oshinko rest server (#28)

* Always rely on the pod environment for the oshinko rest server

The oshinko-get-cluster utility can take an optional "-server" argument
to specify the oshinko rest server which is useful for testing or
running outside of a pod.  However, start.sh will always run in
a pod and should always scan the pod env to discover the rest
server, therefore this optional flag can be removed from start.sh

* Remove "-server" on ephemeral cluster delete as well

* Expose the ephemeral cluster delete option in the templates (#30)

* Expose the ephemeral cluster delete option in the templates

The oshinko app launcher supports deleting ephemeral clusters,
but the option was not exposed in the templates. This option
has been added and default to "yes". If a cluster is created
on-demand for an app, the cluster will be deleted when the app
completes unless the value is changed.

* Set the ephemeral parameter to be required

This prevents the user from erasing the default "yes" and
leaving the box blank. Changes to the start.sh script will
interpret an unset env var as "yes", so make the user
be explicit.

* Change the affirmative value to "true" from "yes"

* setup dockerfile to allow auto-builds in docker hub (#35)

* reorganize the Dockerfile to enable auto-builds

the dockerfile has to install golang to build the oshinko cluster util

* fix typo in oshinko path

* leave WORKDIR untouched

* group chown/chmod

* include missing -R

* update targets that run docker build for new Dockerfile location

* enable travis-ci.org (#36)

* reorganize the Dockerfile to enable auto-builds

the dockerfile has to install golang to build the oshinko cluster util

* fix typo in oshinko path

* leave WORKDIR untouched

* group chown/chmod

* include missing -R

* update targets that run docker build for new Dockerfile location

* enable travis-ci.org

* Move the random cluster name generation into the launcher. (#32)

Rather than generate the cluster name in the templates,
move the random generation into the start.sh script and
allow the cluster name to be blank in the templates.

This allows random cluster name generation to be used
from "oc new-app".

Also, increase the randomized portion to 6 characters.