Fix error handling in oshinko-get-cluster (#38)
* Fix error handling in oshinko-get-cluster

Ignore the auto-generated Error() routines in the oshinko-rest
client which do not correctly print error lists and handle
reporting of error responses locally.

Additionally, ensure that all failures result in os.exit(-1)

* Echo the output from oshinko-get-cluster in an error case

The output of oshinko-get-cluster is stored in an array with
elements separated by line. On error, print a message and
then echo the expansion of the array with lines preserved.