Use a ConfigMap for spark driver configuration
The pyspark templates will mount a ConfigMap on
the driver pod which contains a spark configuration.
By default, it will mount an empty ConfigMap created
by oshinko but the name of another ConfigMap may be
specified in the template. If the start.sh script sees
a non-empty ConfigMap mounted in the container at a
well-known location it will use that spark configuration.