Use the cli to dynamically retrieve driver spark config
This change uses an extended oshinko CLI (oshinko-clix) to
check for the existence of a specified configmap at application
launch time. If this configmap exists, the files it contains
will be written to $SPARK_HOME/conf to update the spark
configuration of the driver.

This replaces an impl that used a volume mount to modify the
spark driver config.