Pull a specific tag of the oshinko-cli when building
This change causes the s2i images to pull a specific
tagged version of the oshinko-cli source code when building
the image. Previously the tip of master would always be
pulled. Adding the tag also gives us a way to trigger
a build of the s2i images with a git commit.