Modify generate_container_user for all S2I images
This change moves generate_container_user to the common directory
and references it in the Dockerfiles for each S2I image. Previously,
pyspark inherited the script from the base image while java and scala
carried separate local versions.

This change also adds the current user to /opt/app-root/etc/passwd
if the user does not exist in /etc/passwd already. The test here is
different than the original -- formerly the user was added only if
it was not 1001 or 0.