Fix signal handling in start.sh
Use exec in each of the s2i/bin/run scripts so that
start.sh runs as PID 1. Additionally, run spark-submit
as a background process so that start.sh can catch a
termination signal during the run of the spark application.
If we receive a signal and we've saved the PID of a background
process, send a signal to that process.