Report when a shared cluster is created instead of ephemeral
Ephemeral clusters are currently only supported for driver pods
that are launched as part of a deployment. Report that a shared
cluster was created instead in cases where the driver tries to
create an ephemeral cluster but is not part of a deployment
(for example it was created via a job).

* also change mode on test/cmd/cluster.sh
* also change "shared" to "<shared>"