This change adds tests for templates maintained in the oshinko-s2i reposoitory as well as https://radanalytics.io/resources.yaml.
These tests are focused on ensuring that the templates are able
to correctly:

* set environment variables expected by the s2i build process
* set environment variables expected by the start.sh script
  in the running application image
* configure downward api elements expected by start.sh

They are not concerned with the actual operation of start.sh
beyond checking that the application pod actually starts
(start.sh functionality is covered elsewhere). In essence, they
are checks of the "API" for our s2i scripts and start.sh.