Include build directories for scala, java, and pyspark
This change includes docker context directories for building
the s2i images. These directories are generated with dogen. The tarball
files have been changed to zero length, and the scripts modified to
download the files if they are zero length. This allows the same
environment to be used with github and other platforms where there
are not resrictions on storing the tarballs.