Update make files and hooks/pre-commit
This change makes it easier to regenerate the build directories,
set the build artifact tarballs to zero length, and extends the
pre-commit hook. Pre-commit now looks for cases where image yaml files
or scripts have changed but the build dirs have apparently not.
This is not an exhaustive, sophisticated check but it should catch
cases where someone forgets to regenerate the build dirs.