Adding a way to pass the python dependencies to spark job
If the repository that is being build by the s2i builder contains
file called 'worker-requirements.txt' or just `requirements.txt` (in which the python deps are
listed), it installes them locally to a directory using pip and bundle
them all together into one big zip file. This zip file is them submitted
together with spark jop using the '--py-files dependencies.zip' option.

note1: by default script looks for worker-requirements.txt and falls back
to requirements.txt. Also if the dependencies.zip file is present
(either it was already in the repo or the container was stopped and
started again), we don't build the dependencies again, but use the
existing zip file instead.

note2: the idea of worker-requrements.txt is to isolate those dependencies that
are necessary only for spark workers, because not all the dependencies driver
needs are also needed on the worker nodes.