Make tests of env vars wait for test app start
This change will make the tests for env var settings wait
for the app to start, and then check the log for any error
output from bash (which will show up as a mention of start.sh).
This will catch bash programming errors that may be dependent
on env var settings.