Remove tini from s2i install
This includes fixing up bootstrap.sh to use a signal handler
and subprocess wait in non-s2i command cases. It also includes
an additional sed modification to the spark entrypoint to remove
a tini call (this is to support s2i with spark-on-k8s where the
built image is also used as the spark cluster image).