Add rsync to the image*-inc.yaml files and regenerate
Also, maintain the image*-inc.yaml files as part of the repository.
They had been constructed on the fly and were not permanent files,
but with the addition of rsync they should be permanent. In the
future maybe differences can be covered with overrides. (They were
actually added to the repository in a recent commit accidentally,
but this turned out to be a good idea).